requirements.txt
pyelftools==0.29
pyserial==3.5
